TVB-2511: Make id's of documentation strings have underscores while keeping the spaces in the displayed model names + moved matjax file.
