+++ framework_tvb/tvb/adapters/simulator/model_forms.py
@@ -62,24 +62,24 @@ def get_model_to_form_dict():
 
 def get_ui_name_to_model():
     ui_name_to_model = {
-        'Generic_2d_Oscillator': ModelsEnum.GENERIC_2D_OSCILLATOR.get_class(),
-        'Kuramoto_Oscillator': ModelsEnum.KURAMOTO.get_class(),
+        'Generic 2d Oscillator': ModelsEnum.GENERIC_2D_OSCILLATOR.get_class(),
+        'Kuramoto Oscillator': ModelsEnum.KURAMOTO.get_class(),
         'supHopf': ModelsEnum.KURAMOTO.get_class(),
         'Hopfield': ModelsEnum.HOPFIELD.get_class(),
         'Epileptor': ModelsEnum.EPILEPTOR.get_class(),
         'Epileptor2D': ModelsEnum.EPILEPTOR_2D.get_class(),
-        'Epileptor_codim_3': ModelsEnum.EPILEPTOR_CODIM_3.get_class(),
-        'Epileptor_codim_3_ultra-slow_modulations': ModelsEnum.EPILEPTOR_CODIM_3_SLOW.get_class(),
+        'Epileptor codim 3': ModelsEnum.EPILEPTOR_CODIM_3.get_class(),
+        'Epileptor codim 3 ultra-slow_modulations': ModelsEnum.EPILEPTOR_CODIM_3_SLOW.get_class(),
         'EpileptorRestingState': ModelsEnum.EPILEPTOR_RS.get_class(),
         'Jansen-Rit': ModelsEnum.JANSEN_RIT.get_class(),
         'Zetterberg-Jansen': ModelsEnum.ZETTERBERG_JANSEN.get_class(),
-        'Reduced_Wong-Wang': ModelsEnum.REDUCED_WONG_WANG.get_class(),
-        'Reduced_Wong-Wang_with_Excitatory_and_Inhibitory_Coupled_Populations': ModelsEnum.REDUCED_WONG_WANG_EXCH_INH.get_class(),
-        'Stefanescu-Jirsa_2D': ModelsEnum.REDUCED_SET_FITZ_HUGH_NAGUMO.get_class(),
-        'Stefanescu-Jirsa_3D': ModelsEnum.REDUCED_SET_HINDMARSH_ROSE.get_class(),
-        'Zerlaut_adaptation_first_order': ModelsEnum.ZERLAUT_FIRST_ORDER.get_class(),
-        'Zerlaut_adaptation_second_order': ModelsEnum.ZERLAUT_SECOND_ORDER.get_class(),
-        'Linear_model': ModelsEnum.LINEAR.get_class(),
+        'Reduced Wong-Wang': ModelsEnum.REDUCED_WONG_WANG.get_class(),
+        'Reduced Wong-Wang with Excitatory and Inhibitory Coupled Populations': ModelsEnum.REDUCED_WONG_WANG_EXCH_INH.get_class(),
+        'Stefanescu-Jirsa 2D': ModelsEnum.REDUCED_SET_FITZ_HUGH_NAGUMO.get_class(),
+        'Stefanescu-Jirsa 3D': ModelsEnum.REDUCED_SET_HINDMARSH_ROSE.get_class(),
+        'Zerlaut adaptation first order': ModelsEnum.ZERLAUT_FIRST_ORDER.get_class(),
+        'Zerlaut adaptation second order': ModelsEnum.ZERLAUT_SECOND_ORDER.get_class(),
+        'Linear model': ModelsEnum.LINEAR.get_class(),
         'Wilson-Cowan': ModelsEnum.WILSON_COWAN.get_class(),
         'Larter-Breakspear': ModelsEnum.LARTER_BREAKSPEAR.get_class()
     }
